第一次提交作业

+++ 02-MyBatis-07-Log/src/main/java/com/fc/entity/Student.java
@@ -8,7 +8,7 @@ import java.util.Date;
 public class Student {
     private Integer id;
     private String name;
-    private Byte age;
+    private Integer age;
     private String gender;
     private Date birthday;
     private String info;
